storage: improve fallocate sizing
When writing past segment size, fallocate just enough to support next
write. After that segment should be rolled anyway.
When very low on disk, still fallocate enough to support the next write.

This does not reduce the number of fallocate syscalls. We just reduce
the size we fallocate for the last batch appended to the segment in the
general case. Hoping that file systems will appreciate this.