serde/parquet: introduce metadata encoding
As part of the journey to write parquet files, we need to be able to
serialize parquet metadata using the Apache Thrift compact wire
protocol. This commit adds some helper classes for writing this format
and then code to serialize parquet file metadata and page headers to
bytes.

Tests will come in a later commit