features: helper to check if major version upgrade
This introduces the `release_version` which is going to be useful to
define which logical version bumps correspond to a major version
release.

This allows us to check if the node is trying to perform a major version
upgrade. In the next commit we use this to perform license validation
only on feature version increases that are major version upgrades.