c/partition: introduced partition property stm
Introduced `cluster::partition_property_stm` the STM is intended to
track partition properties that updates must be strictly ordered against
the partition content.

One of the example of such property is a `writes_disabled` property.
After the property has been set the partition should no longer accept
data batches replicate requests. If that property state would be stored
anywhere else than the partition itself it may always be out of date
on the current partition leader. Hence the property update must be
stored in the data partition itself.

The `partition_properties_stm` update commands contribute to the offset
translation as they are not exposed to Kafka protocol consumers.

Signed-off-by: Michał Maślanka <michal@redpanda.com>