`storage`: add `segment_index::may_have_tombstone_records()` functions
For future changes for sliding window compaction scheduling, it is helpful
to have a flag that indicates whether a segment may contain tombstone records
or not.

Add a new `bool` field, `may_have_tombstone_records` to the `index_state`, and
mark/unmark its value during segment deduplication and data copying.

This field is `true` by default, which can lead to false-positives. A segment
is only considered to not have tombstone records after proven by de-duplication/
segment data copying in the compaction process.