datalake: fix decimal test
The parameters to the decimal type were incorrect in the test code.
Precision must be greater than scale, but it was reversed before. It
didn't matter because we weren't validating that we could translate the
data to Parquet, but once we start translating data to Parquet, this
will generate an error.