iceberg: make snapshot default copyable
It will be helpful to be able to apply table metadata updates that
require a new snapshot to be added to the table (e.g.
table_update::add_snapshot). In such cases, we will use the in-memory
update to apply the update to in-memory table metadata, and also to
serialize a request to send to the catalog. Since there are two usages
of the added snapshot, it isn't sufficient to pass ownership of the
snapshot around.

So, this commit swaps out the snapshot's chunked set for a copyable one.