tests: persistent cache of github releases
The previous file-based cache was only used during a single-run of
ducktape and not persisted on the developer's file system.

This means that we can quickly hit github's rate limiting when iterating
on a ducktape test locally.

To avoid hitting this rate limit, the cache file is mounted from the
developer's (or CI) file system to the ducktape container's file system
to persist it between runs.

(cherry picked from commit e78099a6952213830735273f430a64be82d0a727)