schema_registry/json: introduce json_id_uri as a canonical schema base uri type
the base uri for a schema is saved with the "$id" or "id" key and any
relative reference is relative to this uri.

since the uri can contain a bunch of other parts like protocol (http vs
https) or port, this type is meant to be a canonical representation as
far as references are concerned.

the form is

host[/path]

where /path is optional

new type alias: id_to_schema_pointer

it's a map to resolve references json_id_uri->{json_pointer, dialect}

with a json_id_uri, we get the path to the actual json object for the
bundled schema, and the dialect used by it.

an absolute reference will first query the json_id_uri, get the path to
the root object, and then reach for the specific object