debug_bundle/test: Busy wait for terminated process status
Process status is set in a background fiber that is concurrent to
the test case. So we'll busy wait with a very generous (30s) timeout
at a 200ms interval.

If this test blows the 30s timeout, that's probably indicative of
a real bug in app code and should be addressed in due course. This
commit seeks only to address intermittent test failures arising from
the the common case race condition.

Fixes [CORE-7678]

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>