protobuf: actually output map types
Maps are weird in proto3. Internally they are a repeated synthetic message type
with a key and value, because we're not checking if the type is a map,
we're outputting those as messages. Fix that by when we complete a
message see if it's a map entry then convert it to the proper internal
map type.