iceberg/rest-client: Add a retry policy
A retry policy contains logic for deciding whether an http call should
and can be retried.

It accepts a resolved future returning a collected response (a collected
response is one whose status has been read).

If the future failed, then the exception is used to decide if the failed
attempt can be retried.

If the future is available but the status code is not of the 2xx
variety, another decision is made if the call can be retried.

Finally if the future is successful and the call is also deemed
successful, no retries are necessary.

This is largely modelled after s3/abs retries but the intent is to make
the logic isolated to make testing easier, so that a retry policy can be
passed in to another section of code.