iceberg/rest-client: Add basic catalog client
The catalog client exposes a high level set of operations over the
catalog API. In this changeset a single method is added which acquires
tokens, and a set of helper methods which are expected to be used with
all API calls.

The pattern to follow with all methods of the catalog client should be
similar:

* ensure token exists.
* serialize arguments to json (if payload is required)
* perform call to rest endpoints with optional payload
* deserialize response to domain object

The methods should use expected<> as return type to support chaining
together calls and returning early error in case of failure.