rptest: fix typo in segment readers metric name
We get test failures due to the following assert in the test:

```
assert stats['segment_readers']
```

Based on the comment and the expectations it is clear that the author
intended to use the segment readers metric rather than remote partition
readers metric. Remote partition readers are not even cached so at the
end of the test it is expected not to have any active.

Fixes CORE-7014