application: use contruct_service to build rpc server
This way if there is an error in start()-ing the server (eg: port
conflict), it is cleaned up on exit.

This prevents a crash and gives a meaningful error message when we try
to run a fixture unit test while a redpanda broker is using the same
port in the background, the assert is really confusing and may often be
misinterpreted as a bug in local dev work resulting in a huge waste of
brain cycles.