iceberg: add merge append action
Implements a new merge_append action that adds a given group of data
files to the table, potentially merging them to an existing manifest if
there are too many manifests.

This resembles the implementation in the pyiceberg library[1]: we
bin-pack manifest_files into groups of 8MiB and merge all bins but the
latest one. If the latest bin contains fewer than 100 manifests, it is
left alone, otherwise it is also merged. In the future, these will be
configurable, but for now, this commit adds these default parameters.

Unlike the python implementation, which groups together merges based on
partition spec, this implementation throws if there is more than one
partition spec. This is left as short-term future work.

[1] https://github.com/apache/iceberg-python/blob/e5a58b34dd830c6ffea11649613b693f70f7cbb4/pyiceberg/table/update/snapshot.py#L475