rptest: do not expect cached segment readers at the end of the test
This assertion fails from time to time. It seems perfectly fine there
not to be any cached readers available.

If there aren't any running queries using readers and we are under
memory pressure the next query will likely evict all segment readers to
free up resources. Reader eviction (and memory accounting) is
asynchronous but the metric/probe update is synchronous.

If we get metrics in-between last query acquiring memory resources (or
slightly after) and before the actual physical eviction happened then we
expect the metric to be 0.

It is possible for the last query to be aborted altogether in-between
requesting resources and actually creating a reader which would also
lead to metric being 0.

I consider this acceptable behavior so removing the assert.