rptest: produce more data in FullDiskReclaimTest to trigger gc conditions
For reasons not well understood each partition needs to have at least 2
segments for eviction to be triggered. We have tried to change this
condition in https://github.com/redpanda-data/redpanda/pull/22686 but
had to revert it in
https://github.com/redpanda-data/redpanda/pull/22995.

For now, produce more data to increase the chance of each partition
having at least 2 segments worth of data.