iceberg: botch transactions on first error
Keeps track of the first error seen by a transaction, and subsequent
fails all actions that are performed.

I considered making the table and update getters also return an error,
but I think it makes the code very unwieldy.

Longer term, when the transaction is attempted to be committed, we
should consider the error and fail to commit. That will come in a later
commit.