pp/httpd: Remove extra decode on path param get
seastar::http::request::get_path_param performs URL decoding
automatically. When this path was updated to use the new method,
the pre-existing URL decode behavior was left in place.

This is usually not noticeable because decoding an un-encoded
string is a no-op. However, that's not the case if the decoded
string itself contains URL escape sequences.

For example, given the path parameter "foo%252Fbar":
  - decoded once:   "foo%2Fbar"
  - decoded twice:  "foo/bar"

So this commit removes the "extra" decode step from the PP httpd
module and adds a ducktape test to confirm that URL escape sequences
are preserved in a client-provided subject name, the specific case
that led to this fix.

Unfortunately, get_path_param destroys some information such that
distinguishing between an absent parameter and an ill-formed parameter
is no longer possible. So this commit includes a bit of extra logic
for the nominally unlikely case where get_path_param returns an empty
string.

CORE-7738

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>
(cherry picked from commit bcab3ff93038c059650cd5198a22676f5ba6718a)