iceberg: add filesystem catalog
Adds a very simple implementation of a file system catalog, meant for
operating on Iceberg tables only using object storage. This closely
resembles the concept of "file system tables"[1] (though
recently deprecated by the Iceberg spec). This will still be useful for
test.

There are some caveats called out in the header around concurrency.
Without additional polish, use of the catalog is unsafe and not
recommended for production.

[1] https://iceberg.apache.org/spec/#file-system-tables