ct: Add batcher component
The batcher is a sharded service that accumulates write requests,
aggregates their content and uploads L0 objects in the background. The
batcher propagates placeholder batches back to callers. The caller is
supposed to replace set of original record batches (represented as
model::record_batch_reader) with the set of placeholders.

The batcher implements 'write_and_debounce' method which does all this.
THe caller in the write path invokes this method and passes its reader
that has data from the produce request. The method returns the set of
placeholders for these batches. The placeholders are pointing to the
data in the cloud storage.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>