iceberg: add table_io
Adds another wrapper around cloud_io::remote, but this time for table
metadata. The intent here is to be used for a filesystem-like catalog
that will use S3 for table metadata management instead of a catalog
service like Polaris.

To that end, we'll need to be able to read/write table metadata (in this
commit) and version hints (in a later commit).

I considered merging this with manifest_io but opted not to since we'll
only need table_io for the filesystem catalog -- for other catalogs,
table metadata will be managed by a catalog service.