rptest: increase threshold for test_full_disk_triggers_gc
Another take on fixing
rptest.tests.full_disk_test.FullDiskReclaimTest.test_full_disk_triggers_gc.

Previous one was https://github.com/redpanda-data/redpanda/pull/23512
but I found another case where it fails. It is possible that after
eviction a leadership transfer or election will happen thus creating a
new log segment...

We can't wait for that segment to be eventually evicted because we
currently do not evict last segment. We could disable leadership
balancer but there might be a network issue causing a leader election in
a CDT test. It is hard to impossible to account for all conditions.

Instead, change the thresholds on the test such that it is acceptable to
have at least one segment still on disk per partition.

While it is possible for this test to pass with a partition having 5
segments and others having none, I don't think that's something this
test needs to worry about.