datalake: added support for conversion from protobuf to iceberg value
Added a support of converting a Protobuf deserialized message to
`iceberg::value` using its `Descriptor`. The conversion is being done by
traversing the schema stored in the `Descriptor` as the desrialized
representation does not contain default values even when they are
present. The conversion algorithm generates a default value for the
field if it is not present in parsed representation.

Signed-off-by: Michał Maślanka <michal@redpanda.com>