`kafka`: add `hide_disabled_tristate()`
Similar in intention to the `override_if_not_default` overload for
`std::optional<>`. This ensures that `tristate` topic properties may
still display `DEFAULT_CONFIG` instead of `DYNAMIC_TOPIC_CONFIG` in the
case that their value is `{disabled}` and the cluster default is `std::nullopt`.

This is useful for topic properties that may want to default to a `{disabled}`
state without indicating `DYNAMIC_TOPIC_CONFIG` as a source.