iceberg: version_hint io helpers
Adds helpers to table_io for dealing with version_hint files. The
version hint is a file used in the Java implementation's
HadoopTableOperations[1], which is typically what writes table metadata
to S3.

The file contains an stringified integer N, and is meant to be used to
determine the table metadata path: <table_location>/v<N>.metadata.json.

[1] https://github.com/apache/iceberg/blob/c862b9177af8e2d83122220764a056f3b96fd00c/core/src/main/java/org/apache/iceberg/hadoop/HadoopTableOperations.java