c/feature_manager: add fallback license env var
This adds an environment variable that can be used to insert an
enterprise license on startup into a non-compliant cluster when it is
failing the license enforcement check. This should be an undocumented
escape hatch we can use if we ever need to inject a license into this
check without being able to add it into the controller log via
`rpk cluster license set` (eg. if we can't downgrade the broker to a
previous version).