k/group: do not ignore an error when aborting expired transaction
When transaction is expired due to a timeout a group coordinator issues
a request to a transaction coordinator to abort the transaction. The
operation may fail as the currently known coordinator may not be a
leader or the request may timeout.
The result of transaction abort operation was previously ignored. This
lead to a situation in which an error when aborting transaction lead to
a tight loop in group coordinator as it instantly dispatched another
abort retry. With the fix the error when aborting ongoing transaction is
not ignored but rather than that if an error happened it is logged with
warn severity and the coordinator waits for the next scheduled timer
ticked (by default ten seconds) before retrying.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 00fb835cd7883547616808ce6f508249fb55deb9)