ct: Add basic_cache_service_api interface
The interface is extracted from the 'cloud_storage::cache'. The idea is
that the code that uses the service doesn't have to depend on
'cloud_storage' which is a big dependency.

The implementation will not be provided here in 'cloud_io'. Instead the
actual 'cloud_storage::cache' will inherit this interface. The interface
is a template which can be parameterized by clock. This is needed to
allow testing with manual clock. The 'cloud_storage::cache' is supposed
to be derived from the 'basic_cache_service_api<ss::lowres_clock>'.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>