ct: Derive 'cloud_storage::cache' from the i-face
The cache implements 'basic_cache_service_api<ss::lowres_clock>' now.
Some utility types are aliases for 'cloud_io' types now. This change
doesn't change the functioning of the cache. It's only introduces the
interface so the code inside the 'cloud_topics' could depend on it
instead of the 'cloud_storage' (which is supposed to be retiered
eventually).

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>