debug_bundle/admin: Added expired status
The purpose of the expired status is to inform the user that a debug
bundle was at one time generated, but has been cleaned up by the process
controlled by the debug_bundle_auto_removal_seconds cluster
configuration.

Added a new error code returned when attempting to get the debug bundle
file to indicate that it was cleaned up (debug_bundle_expired).

Updated the admin API handlers to return 410 - GONE if
debug_bundle_expired is returned by the service.

Signed-off-by: Michael Boquard <michael@redpanda.com>