cst/inv/ducktape: Remove inventory report from test
The test became flaky since adding an inventory report. The cause of the
flakiness is leadership transfer during test and the constraint we apply
that one report is only processed by a node once. Consider the following
scenario:

* Node A leaders partition p1. Node B leads some other partitions.
* Both nodes process the latest report R.
* p1 moves leadership to B. Now despite the test having a short check
  interval for the report, we only create R once in the test, so B will
  never process R again.
* On the next scrub, B will not have the data for p1, because when R was
  processed, rows for p1 were skipped by B, as it did not lead p1 then.
* As data is missing, the scrub will not check for the segment
  existence, thus not finding anomaly and breaking the test.

The solution is for inv. service on each node to maintain a delta of
partition leaderships. If we gain new partitions then we should process
the report again for those partitions, so we can next scrub with latest
data for the partitions.

This is a mechanically trivial change but will require some time to
implement, to avoid the flakiness right now the test is reverted to a
state before the changes related to inventory report were introduced.

Once the delta related changes are in place the test can be extended
again to cover inventory based scenario.