tests: test legacy dashboard in rpk generate
The legacy dashboard uses the admin API to get the
metrics and auto-generate a dashboard.

Also, removes the Title assertion as this depends
on an external repo. We just need to test if rpk
is able to download and parse.

(cherry picked from commit ebe87b95ed5ca12881f96cf62827c00514883b6a)