features: delay first license nag
If a customer bootstraps a cluster and immediately installs a license we
still print the license nag today. This is confusing to compliant
customers (eg. RP Cloud).

This change delays the license nag by 1 interval.

Note that this means that a savvy customer could increase the license
nag interval to infinity and thereby disable the nag. But that seems
acceptable given that this is just a log message.