schema_registry/json: reduce the scope of get_object_or_empty
before this commit, get_object_or_empty would get a context to pass to
get_schema. this was done with the idea of resolving references in
get_schema and benefit from that.

however, this clash with the design of resolving references only in
is_superset, and the standard prevents "properties" "patternProperties"
"dependency" from being references (see tests).

this means that the whole code can be simplified (at the cost of handing
objects and bools directly in get_object_or_empty).

keywords (like items, additionalItems, additionalProperties) that can be
references where already passed directly to is_superset, so resolving
early the references is a capability that's not needed