schema_registry/json: merge_references helper
this function accepts list of json object to merge into a new schema
with the form

{
  "allOf": [ *input_list ]
}

the input_list is expected to be a chain of reference objects of size
>=2, all of them with a $ref key except the last one

the final schema will contain these objects, with the $ref key removed
from each one

the user is responsible to keep the original json::Document alive, since
the non-ref strings will be referenced in the result.

one optimization: if the chain contains only one non-empty object
without the $ref key, this will be returned directly.
this is to support the likely case of a chain of pure $ref pointing to
the final schema