schema_registry/json: resolve references in get_schema
get_schema will use resolve_reference to traverse references until the
final target is reached.

it's expected for references to absolute, thanks to the preprocessing
phase.
retrieval is two step process: first retrieve the bundled schema, then
retrieve the sub object.

if the sub object contains a reference, repeat the process.

the function uses schema_context to limit the number of iterations. the
number of iterations is shared across children of an object to prevent
indirect references to cause a stack overflow.

external references are not supported

different dialects in bundled schemas are not supported in this commit,
the function throws if the dialects do not match