datalake/coordinator: add state
Adds structs that track the datalake state for each topic. At a high
level the state per partition is a FIFO collection of contiguous offset
ranges that have been translated into Parquet.

An upcoming commit will introduce serializable updates that can be
applied to this state to ensure various invariants.