serde/parquet: add level encoding
Before data is encoded v2 data pages, the "levels" are written. These
are the definition and repetition levels. These are used for knowing the
nesting depth of values (needed for knowing what level a null value was)
and the index if the value is a repeated type.

You can read more about levels generally here:
https://blog.x.com/engineering/en_us/a/2013/dremel-made-simple-with-parquet

We choose to implement this as simply as possible for the first
iteration by always choosing run length coding for values. The low bit
in the header is used to determine if the run length (0) or bit packing
is used (1). Hence the reason for `unsigned_vint::to_bytes(length <<
1)`. A more complicated algorithm could be used to alternate between
bitpacking and run length encoding. This is a similar tradeoff to what
DuckDB makes.

Otherwise there are links in the code for pointers to the relevant spec.
The codebases for duckdb and parquet-go were studied for the development
of this feature, and relevant code pointers are as follows:

https://github.com/parquet-go/parquet-go/blob/1b52c8764200633b0bbd093fc9fed9b96ded799d/writer.go#L1066
https://github.com/duckdb/duckdb/blob/123b82b9053c4843559035b6723c867b2618b2d9/extension/parquet/column_writer.cpp#L92