bazel: Enable LTO for redpanda
This patch enables LTO for redpanda by default (main binary only).

We add `-flto=thin` and `-ffat-lto-objects` to the compile flags. This
allows for the final link step to chose whether to use LTO or not (by
passing `flto` to the linker step).

Hence, we can selectively enable LTO only for the redpanda binary and
not for the test binaries. Note this isn't a massive advantage yet for
the bazel build because of the thing where it builds everything twice
anyway). It is however already useful for not using LTO for initial
local dev (lto=False) but then enabling it for an eventual microbench or
a package build.

We chose thin-lto over full lto as compile times are a lot slower with
that (~5x) and additional perf improvements are small. Also we can
always change later.

As is the final link step is about a minute. This will go up once we
enable debug info as that has quite the impact on compile and link
times.

Note we can't use the clang thin-lto cache option because it would
require a bunch of bazel work to use a cache dir.

In a librdkafka bench this change shows a 2-3% improvement. There is
some microbench results in CORE-327 but note that microbenches aren't
really that representative for LTO.

Ref CORE-7789