bazel: Enable LTO for redpanda
We add `-flto` and `-ffat-lto-objects` to the compile flags. This allows
for the final link step to chose whether to use LTO or not (by passing
`flto` to the linker step).

Hence, we can selectively enable LTO only for the redpanda binary and
not for the test binaries. Note this isn't a massive advantage yet for
the bazel build because of the thing where it builds everything twice
anyway).