tx/group: ignore replay requests if producer not found
If a producer/transaction is not found, we assume it is replay request
from the coordinator to roll forward/back the transaction. This is valid
because the the current request to commit/abort confirms that the
coordinator sees a commit/abort respectively in its's log, so the
original request that cleaned up the state should have been the same
too. One particular sequence of events where this is possible is as
follows.

- Coordinator - Attempting to commit -- it got a timeout but the group
	     commit actually made it to replica 1
- group leadership changed from replica 1 to replica 2
- replica 2 has a compacted state of the log
- replica 2 recovers and _producers is empty
- replica 2 now rejects the commit request

Note: compaction cleans up the producer ids in group topic, so the
resulting log after compaction cannot load up the producer ids resulting
in a pid not found.