dt/license: fix flaky license enforcement test
The role creation was racing with the node restart. If the target node
is restarted before it observes the new role, it won't crash on startup
like we would expect it to.

To fix this, we want to ensure that the enabled enterprise feature is
observable on all nodes before the restart. We could implement this with
roles, but it is simpler to switch to a different enterprise feature and
using `redpanda.set_cluster_config`, which implements
all-node-observability under the hood.