`kafka`: fixes and interface change to `to_cluster_type()`
This commit changes the behavior of `to_cluster_type()`, as well
as the `create_topics` handler as a whole.

Before, if any invalid topic configurations in a `create_topics_request`
lead to a function called from `to_cluster_type()` throwing an exception,
the request would continually be retried until a timeout was hit.

This had the effect of "all-or-nothing" behavior with topic creation,
in which just one invalid topic config would result in none of the topics
being created.

By catching and handling the exception in `to_cluster_type()`, we now change
the behavior to allow for partial creation of topics, in which those topics
with valid configurations are succesfully created, while the invalid topics
have their errors logged and returned.

An additional iterator is passed in order to generate any necessary
errors in the `create_topics_response`, which is returned back to the
`kafka::request_context`.