datalake/coordinator: iron out RPC types
Updates the coordinator RPC add_file to use translated_offset_range to
represent groups of files. Each translated offset range will be
analagous to the write of a single translation operation, as it
corresponds to a single contiguous Kafka offset range.

Also updates the fetch_lastest to just return the latest Kafka offset.