`kafka`: add `sort_topic_response()` in `create_topics.cc`
In the case that some topic configurations were invalid while others
were valid, the `create_topics_response` vector may have had results
pushed back to it in an order that is different than the order in
which the topics were supplied in the request.

Add a new function `sort_topic_response()` that sorts the response's
`topics` vector based on the order of the `create_topics_request`
provided before returning it.