c/migrations: retry advance requests
When migrations backend has ensured all the work for a migration has been done
it sends a request to advance the migration to the new stage. This request may
fail, in particular when controller leader changes. Previously they were only
retried when the node becomes leader and therefore migration coordinator again.
This has two problems:
1) The node may lose leadership only briefly, so that it never ceases to be the
migration coordinator. We might solve it by always retrying advance requests on
leadership notifications.
2) The request may fail for another reason, such as a timeout.

So we retry them indefinitely with 5s sleep. We only stop doing so if we receive
a successful reply from migration frontend, or if the migration is known to have
advanced, whether as a result of our actions or not.