dl/coordinator: add coordinator manager
Since the coordinator is going to reference cloud IO, the Iceberg
catalog, etc., we'll generally want to prefer managing its lifecycle
outside of cluster::partition or the STM or anything like that. Despite
the underlying STM being managed by the partition, the
non-deterministic, external dependencies used by the coordinator make it
much less desirable to tie so deeply into our core.

So, this adds a coordinator manager that will be managed at the
application layer, that will take care of managing the coordinator's
lifecycle. This is similar to the new archiver_manager, in that it hooks
into all the same manage/unmanage/leadership hooks, constructing
long-lived coordinators for as long as the partition is on the given
shard.