ct: Add 'overlay_collection' component
The overlay_collection is a collection of overlay batches that uses
'patience sort' algorithm. The idea here is that the collection is based
on a series of sorted runs which can be compacted using k-way merge. The
sorted run can potentially be stored using the columnar format which
will allow us to increase metadata density per partition.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>