c/tx_gateway: fixed missing set outcome value after tx is committed
When a transaction is already committed and the client is retrying the
request a coordinator may find the transaction outcome already
available. In this case the client should receive the result that is
known already instead of an error. In the previous code the outcome was
not set which resulted in `unknown_server_error` being returned to the
client. This in the end caused a chaos tests to fail.

Signed-off-by: Michał Maślanka <michal@redpanda.com>