datalake: add local_data_file to store local paths
This separate the concepts of a local file from a remote file. For
local files it splits path into two parts, base_path and file_path.
The uploader will not include base_path in the path for the remote
file. This also modifies the writers to use local_data_file instead
of data_file.