dl/coordinator: use STM shared ptr in coordinator
The upcoming ownership model for the coordinator will be:

The STM:
- partition_manager owns partition
- partition owns log
- log owns STM manager and coordinator_stm

The coordinator:
- coordintor_manager owns coordinator
- coordinator refers to coordinator_stm, ultimately owned by the
  partition, per above

The coordinator_manager will respond to events like the partition being
managed/unmanaged, but ultimately needs to perform synchronous
operations when these events happen. Stopping a coordinator will
necessarily be asynchronous though, since it needs to wait for in-flight
operations to finish.

So, this makes the coordinator take a shared_ptr to the STM, so that it
can background the stopping of the coordinator.