dl/coordinator: add file committer
Adds an class to be used by the coordinator to encapsulate committing
files to the Iceberg catalog. It takes as inputs the STM state for a
given topic, and returns the resulting STM updates that capture the
catalog updates it has performed.

The separate abstraction isn't strictly required, but it does make it
easier to test, and helps avoid making the coordinator control loop very
bloated.