datalake: cloud uploader cleans up files
This adds cleanup methods to the cloud uploader. When it successfully
uploads all data files, it cleans up the local temporary files. If
the upload fails, it deletes both the local files and whatever remote
files were uploaded. The coordinator, operating at a higher level, is
repsonsible for retrying all of this.