partition_mgr/notification: notify after start
A managed notification currently happens before the partition boostraps.
A bug surfaced in datalake/coordinator_mgr where it was trying to access
stm state as a part of notification which resulted in a segfault because
the state machines haven't been initialized by the time notification
triggered.

This fix waits until the partition boostraps to trigger a managed
notification so the waiters have access to stm state. Judging by the
current users of this notification, it doesn't seem the timing of
notification (whether it happens before the boostrap) is important.