errors: Add 'maybe' template
The 'maybe' template is a utility which is quite similar to
boost::outcome. The difference is that it doesn't depend on
std::error_code and can be used with enum instead.

The header also specializes coroutine_traits_base template
for this type so the coroutines that return
ss::future<maybe<T, E>> are handled by this specialization.

The specialization is converting all exceptions to the error type used
in 'maybe' type parameter. In order for this to work the user of the
template has to specialize 'exception_mapping_policy' template for the
custom error type used in 'maybe'. The policy takes std::exception_ptr
and returns an error code.

The conversion from exception to maybe type can only happen in
coroutines. The user can still use ss::make_exception_future<maybe<T,
E>> function. The problem is that C++ doesn't allow us to specialize
function templates and this function can't be overloaded properly. The
solution is to write a custom clang-tidy rule that forbids usage of
'make_exception_future<maybe<T, E>>' in the codebase.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>