schema_registry/json: add the support to resolve external references
this is done with schema_context::resolve_reference. the function will
get an uri and return a json::Value::ConstObject or throw.

the search is done with schema_index, a local_ptr will be resolved
against the root schema and an external_ptr will resolve against one of
the external_schemas.

note that there is no behavior change yet, because the external_schemas
map is not populated yet