schema_registry/json: move parse_json to jsoncons-only
this is done to reduce conversions when parse_json will call itself to
handle external references.

the name change to parse_jsoncons, and a new parse_json calls the former
and applies the translation to rapidjson of the result.

collect_bundled_schema_and_fix_ref has to change too.

no behavior change yet