schema_registry/json: parse_json extract external schemas
this commit adds support for external schemas: parse_json gets an
(optional) id for the root schema, and a list of external schemas name
to build into the result.

each external schema il collected, parse_jsoncons invoked on it, and
it's list of bundles schemas is added to the root schemas_index.

each bundled schema has to have an unique id (due to the current
resolve_reference implementation) and the external_schema.$id has to be
unique (if it's not present, the exernal schema name will be used as
$id)